More logging fixes and some cleanup
