+++ src/main/kotlin/com/dprint/services/FormatterService.kt
@@ -1,7 +1,7 @@
 package com.dprint.services
 
 import com.dprint.core.Bundle
-import com.dprint.messages.DprintMessage
+import com.dprint.core.LogUtils
 import com.dprint.services.editorservice.EditorServiceManager
 import com.dprint.services.editorservice.FormatResult
 import com.intellij.openapi.application.ReadAction
@@ -61,7 +61,11 @@ class FormatterService(private val project: Project) {
                 try {
                     val editorServiceInstance = editorServiceManager.maybeGetEditorService()
                     if (editorServiceInstance == null) {
-                        LOGGER.info(Bundle.message("formatting.service.editor.service.uninitialized"))
+                        LogUtils.info(
+                            Bundle.message("formatting.service.editor.service.uninitialized"),
+                            project,
+                            LOGGER
+                        )
                         return
                     }
 
@@ -78,11 +82,7 @@ class FormatterService(private val project: Project) {
 
                         val result = resultFuture.get(FORMATTING_TIMEOUT_SECONDS, TimeUnit.SECONDS)
 
-                        result.error?.let {
-                            LOGGER.info(Bundle.message("logging.format.failed", filePath, it))
-                            project.messageBus.syncPublisher(DprintMessage.DPRINT_MESSAGE_TOPIC).info(it)
-                        }
-
+                        // Error logging is handled in the EditorService
                         result.formattedContent?.let {
                             WriteCommandAction.runWriteCommandAction(project) {
                                 getDocument(project, virtualFile)?.setText(it)
@@ -102,9 +102,7 @@ class FormatterService(private val project: Project) {
 
     private fun handleFormatException(e: Exception) {
         // In the event that the editor service times out we restart
-        LOGGER.error(Bundle.message("error.dprint.failed"), e)
-        project.messageBus.syncPublisher(DprintMessage.DPRINT_MESSAGE_TOPIC)
-            .info(Bundle.message("error.dprint.failed.timeout", FORMATTING_TIMEOUT_SECONDS))
+        LogUtils.error(Bundle.message("error.dprint.failed"), e, project, LOGGER)
         editorServiceManager.restartEditorService()
     }
