+++ src/main/kotlin/com/dprint/services/editorservice/v5/EditorServiceV5.kt
@@ -11,16 +11,13 @@ import com.intellij.openapi.project.Project
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
-import org.apache.lucene.util.ThreadInterruptedException
 import java.util.concurrent.CompletableFuture
+import java.util.concurrent.ExecutionException
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
-import java.util.concurrent.atomic.AtomicInteger
 import kotlin.concurrent.thread
 
 private val LOGGER = logger<EditorServiceV5>()
-private var messageId = AtomicInteger(0)
-private const val SLEEP_TIME = 500L
 private const val FORCE_DESTROY_DELAY = 1000L
 private const val FORMATTING_TIMEOUT_SECONDS = 10L
 
@@ -31,83 +28,21 @@ class EditorServiceV5(val project: Project) : EditorService {
     private var stdoutListener: Thread? = null
     private val pendingMessages = PendingMessages()
 
-    // TODO pull this out into a listener class
-    @Suppress("TooGenericExceptionCaught")
     private fun createStdoutListener(): Thread {
         if (stdoutListener != null) {
             stdoutListener?.interrupt()
         }
-        val runnable = Runnable {
-            LOGGER.info("Dprint: Started listener")
-            while (true) {
-                if (Thread.interrupted()) {
-                    return@Runnable
-                }
-                try {
-                    val messageId = editorProcess.readInt()
-                    val messageType = editorProcess.readInt()
-                    val bodyLength = editorProcess.readInt()
-                    val body = MessageBody(editorProcess.readBuffer(bodyLength))
-                    editorProcess.readAndAssertSuccess()
 
-                    when (messageType) {
-                        MessageType.SuccessResponse.intValue -> {
-                            val responseId = body.readInt()
-                            val result = PendingMessages.Result(MessageType.SuccessResponse, null)
-                            pendingMessages.take(responseId)?.let { it(result) }
-                        }
-                        MessageType.ErrorResponse.intValue -> {
-                            val responseId = body.readInt()
-                            val errorMessage = body.readSizedString()
-                            LOGGER.info(Bundle.message("editor.service.received.error.response", errorMessage))
-                            val result = PendingMessages.Result(MessageType.ErrorResponse, errorMessage)
-                            pendingMessages.take(responseId)?.let { it(result) }
-                        }
-                        MessageType.Active.intValue -> {
-                            sendSuccess(messageId)
-                        }
-                        MessageType.CanFormatResponse.intValue -> {
-                            val responseId = body.readInt()
-                            val canFormatResult = body.readInt()
-                            val result = PendingMessages.Result(MessageType.CanFormatResponse, canFormatResult == 1)
-                            pendingMessages.take(responseId)?.let { it(result) }
-                        }
-                        MessageType.FormatFileResponse.intValue -> {
-                            val responseId = body.readInt()
-                            val hasChanged = body.readInt()
-                            val text = when (hasChanged == 1) {
-                                true -> body.readSizedString()
-                                false -> null
-                            }
-                            val result = PendingMessages.Result(MessageType.FormatFileResponse, text)
-                            pendingMessages.take(responseId)?.let { it(result) }
-                        }
-                        else -> {
-                            val errorMessage = Bundle.message(
-                                "editor.service.unsupported.message.type",
-                                messageType
-                            )
-                            LogUtils.info(errorMessage, project, LOGGER)
-                            sendFailure(messageId, errorMessage)
-                        }
-                    }
-                } catch (e: ThreadInterruptedException) {
-                    LOGGER.info(e)
-                    return@Runnable
-                } catch (e: Exception) {
-                    LogUtils.error(Bundle.message("editor.service.read.failed"), e, project, LOGGER)
-                    Thread.sleep(SLEEP_TIME)
-                }
-            }
-        }
 
         return thread {
-            runnable.run()
+            StdoutListener(editorProcess, pendingMessages).run()
         }
     }
 
     override fun initialiseEditorService() {
-        LogUtils.info("Initializing EditorServiceV5", project, LOGGER)
+        LogUtils.info(
+            Bundle.message("editor.service.initialize", getName()), project, LOGGER
+        )
         dropMessages()
         editorProcess.initialize()
         stdoutListener = createStdoutListener()
@@ -118,7 +53,7 @@ class EditorServiceV5(val project: Project) : EditorService {
     }
 
     override fun destroyEditorService() {
-        LogUtils.info("Destroying EditorServiceV5", project, LOGGER)
+        LogUtils.info(Bundle.message("editor.service.destroy", getName()), project, LOGGER)
         val message = createNewMessage(MessageType.ShutDownProcess)
         editorProcess.writeBuffer(message.build())
         stdoutListener?.interrupt()
@@ -157,7 +92,12 @@ class EditorServiceV5(val project: Project) : EditorService {
                 false
             }
         } catch (e: TimeoutException) {
-            LOGGER.warn(e)
+            LOGGER.error(e)
+            pendingMessages.take(message.id)
+            initialiseEditorService()
+            false
+        } catch (e: ExecutionException) {
+            LOGGER.error(e)
             pendingMessages.take(message.id)
             initialiseEditorService()
             false
@@ -169,11 +109,7 @@ class EditorServiceV5(val project: Project) : EditorService {
     }
 
     override fun fmt(
-        filePath: String,
-        content: String,
-        startIndex: Int?,
-        endIndex: Int?,
-        onFinished: (FormatResult) -> Unit
+        filePath: String, content: String, startIndex: Int?, endIndex: Int?, onFinished: (FormatResult) -> Unit
     ): Int {
         LogUtils.info(Bundle.message("formatting.file", filePath), project, LOGGER)
         val message = createNewMessage(MessageType.FormatFile)
@@ -188,11 +124,11 @@ class EditorServiceV5(val project: Project) : EditorService {
             if (it.type == MessageType.FormatFileResponse && it.data is String?) {
                 formatResult.formattedContent = it.data
             } else if (it.type == MessageType.ErrorResponse && it.data is String) {
-                LogUtils.warn(it.data, project, LOGGER)
+                LogUtils.warn(Bundle.message("editor.service.format.failed", filePath, it.data), project, LOGGER)
                 formatResult.error = it.data
             } else if (it.type != MessageType.Dropped) {
                 val errorMessage = Bundle.message("editor.service.unsupported.message.type", it.type)
-                LogUtils.warn(errorMessage, project, LOGGER)
+                LogUtils.warn(Bundle.message("editor.service.format.failed", filePath, errorMessage), project, LOGGER)
                 formatResult.error = errorMessage
             }
             onFinished(formatResult)
@@ -201,7 +137,7 @@ class EditorServiceV5(val project: Project) : EditorService {
 
         editorProcess.writeBuffer(message.build())
 
-        LogUtils.info("Created formatting task for $filePath with id ${message.id}", project, LOGGER)
+        LogUtils.info(Bundle.message("editor.service.created.formatting.task", filePath, message.id), project, LOGGER)
 
         return message.id
     }
@@ -212,7 +148,7 @@ class EditorServiceV5(val project: Project) : EditorService {
 
     override fun cancelFormat(formatId: Int) {
         val message = createNewMessage(MessageType.CancelFormat)
-        LogUtils.info("Cancelling format $formatId", project, LOGGER)
+        LogUtils.info(Bundle.message("editor.service.cancel.format", formatId), project, LOGGER)
         message.addInt(formatId)
         editorProcess.writeBuffer(message.build())
         pendingMessages.take(formatId)
@@ -220,29 +156,12 @@ class EditorServiceV5(val project: Project) : EditorService {
 
     private fun dropMessages() {
         for (message in pendingMessages.drain()) {
-            LogUtils.info("Clearing message ${message.key}", project, LOGGER)
+            LogUtils.info(Bundle.message("editor.service.clearing.message", message.key), project, LOGGER)
             message.value(PendingMessages.Result(MessageType.Dropped, null))
         }
     }
 
-    private fun sendResponse(message: Message) {
-        editorProcess.writeBuffer(message.build())
-    }
-
-    private fun sendSuccess(messageId: Int) {
-        val message = createNewMessage(MessageType.SuccessResponse)
-        message.addInt(messageId)
-        sendResponse(message)
-    }
-
-    private fun sendFailure(messageId: Int, errorMessage: String) {
-        val message = createNewMessage(MessageType.ErrorResponse)
-        message.addInt(messageId)
-        message.addString(errorMessage)
-        sendResponse(message)
-    }
-
-    private fun createNewMessage(type: MessageType): Message {
-        return Message(messageId.incrementAndGet(), type)
+    private fun getName(): String {
+        return this::class.java.simpleName
     }
 }
