+++ src/main/resources/messages/Bundle.properties
@@ -1,51 +1,59 @@
 name=Dprint
-config.name=Dprint
-config.enable=Enable dprint
-config.reload=Reload
-config.run.on.save=Run dprint formatter on save
-config.dprint.config.location=Dprint configuration json file location
+action.com.dprint.actions.ReformatAction.description=Reformat the currently open file with dprint
+action.com.dprint.actions.ReformatAction.text=Reformat With Dprint
+action.com.dprint.actions.RestartAction.description=Restart the dprint editor-service daemon
+action.com.dprint.actions.RestartAction.text=Restart Dprint
 config.dprint.config.invalid=Invalid config file
-config.dprint.executable.location=Dprint executable location
+config.dprint.config.location=Dprint configuration json file location
+config.dprint.editor.info=Received editor info: {0}
 config.dprint.executable.invalid=Invalid executable
-config.dprint.schemaVersion.older=Your installed version of dprint is out of date. Apologies, but please update to the latest version
+config.dprint.executable.location=Dprint executable location
 config.dprint.schemaVersion.newer=Please upgrade your editor extension to be compatible with the installed version of dprint
 config.dprint.schemaVersion.not.found=Unable to determine a schema version
-config.dprint.editor.info=Received editor info: {0}
-progress.formatting=Formatting with dprint
-error.executable.path=Unable to retrieve a valid dprint executable path.
+config.dprint.schemaVersion.older=Your installed version of dprint is out of date. Apologies, but please update to the latest version
+config.enable=Enable dprint
+config.name=Dprint
+config.reload=Reload
+config.run.on.save=Run dprint formatter on save
+dprint.formatting.on=Running format on {0}.
+editor.process.cannot.get.editor.service.process=Cannot get a running editor service
+editor.service.cancel.format=Cancelling format {0}
+editor.service.clearing.message=Clearing message {0}
+editor.service.created.formatting.task=Created formatting task for {0} with id {1}
+editor.service.destroy=Destroying {0}
+editor.service.format.failed=Dprint failed to format the file {0} due to:\n{1}
+editor.service.incorrect.message.size=Incorrect message size, expected {0} and got {1}
+editor.service.initialize=Initializing {0}
+editor.service.read.failed=Failed to read stdout of the editor service
+editor.service.received.error.response=Received failure message: {0}.
+editor.service.shutting.down=Shutting down current editor service
+editor.service.started.stdout.listener=Started stdout listener
+editor.service.starting.working.dir=Starting editor service with executable {0}, config {1}. No working dir found.
+editor.service.starting=Starting editor service with executable {0}, config {1} and working directory {2}.
+editor.service.unable.to.determine.if.can.format=Unable to determine if the file {0} can be formatted.
+editor.service.unsupported.message.type=Received unsupported message type {0}.
 error.config.path=Unable to retrieve a valid dprint configuration path.
 error.dprint.failed=Dprint failed to process the file.
-error.dprint.failed.timeout=The format command timed out after {0} seconds.
-error.failed.to.parse.json.schema="Failed to parse JSON schema. Schema: {0}. Error: {1}."
-logging.format.failed=Dprint failed to format the file {0} due to:\n{1}
+error.executable.path=Unable to retrieve a valid dprint executable path.
+error.failed.to.parse.json.schema=Failed to parse JSON schema. Schema: {0}. Error: {1}.
+formatting.can.format={0} can be formatted
+formatting.cannot.determine.file.path=Cannot determine file path to format.
+formatting.cannot.format=Cannot format {0}.
+formatting.checking.can.format=Checking if {0} can be formatted.
+formatting.error=Formatting error
+formatting.file=Formatting {0}
+formatting.received.success=Received success bytes.
+formatting.received.value=Received value: {0}
+formatting.sending.success.to.editor.service=Sending success to editor service.
+formatting.sending.to.editor.service=Sending to editor service: {0}.
+formatting.service.editor.service.uninitialized=The editor service has not been initialized
+notification.config.not.found=Could not detect a config file
+notification.executable.not.found=Could not find a dprint executable
 notification.group.name=Dprint
 notification.invalid.config.path=The configured Dprint config file is invalid
 notification.invalid.default.config=Found invalid config file {0}
-notification.config.not.found=Could not detect a config file
 notification.invalid.executable.path=The configured dprint executable is invalid
-notification.executable.not.found=Could not find a dprint executable
-formatting.service.editor.service.uninitialized="The editor service has not been initialized"
-editor.service.shutting.down=Shutting down current editor service
-editor.service.starting=Starting editor service with executable {0}, config {1} and working directory {2}.
-editor.service.starting.working.dir=Starting editor service with executable {0}, config {1}. No working dir found.
-editor.service.unable.to.determine.if.can.format=Unable to determine if the file {0} can be formatted.
-editor.service.unsupported.message.type=Received unsupported message type {0}.
-editor.service.received.error.response=Received failure message: {0}.
-editor.service.read.failed=Failed to read stdout of the editor service
-editor.process.cannot.get.editor.service.process=Cannot get a running editor service
-dprint.formatting.on=Running format on {0}.
+progress.formatting=Formatting with dprint
 reformat.action.run=Running reformat action on {0}.
+restart.action.run=Running restart action
 save.action.run=Running save action on {0}.
-formatting.checking.can.format=Checking if {0} can be formatted.
-formatting.can.format={0} can be formatted
-formatting.cannot.format=Cannot format {0}.
-formatting.cannot.determine.file.path=Cannot determine file path to format.
-formatting.file=Formatting {0}
-formatting.sending.to.editor.service=Sending to editor service: {0}.
-formatting.sending.success.to.editor.service=Sending success to editor service.
-formatting.received.success=Received success bytes.
-formatting.received.value=Received value: {0}
-action.com.dprint.actions.ReformatAction.text=Reformat With Dprint
-action.com.dprint.actions.ReformatAction.description=Reformat the currently open file with dprint
-action.com.dprint.actions.RestartAction.text=Restart Dprint
-action.com.dprint.actions.RestartAction.description=Restart the dprint editor-service daemon
